Full pass over BSDF interface and implementations (#73)
* Full pass over BSDF interface and implementations

- Significant improvements to generic microfacet (GGX/Beckmann)
  sampling and evaluation routines. The arithmetic is simpler,
  numeric robustness has been improved, and nonlinear iterative
  solves now use simpler logic and a smaller number of iterations.

- Simplified Fresnel functions for dielectrics and conductors

- Added polarization-aware versions of Fresnel functions for
  dielectrics and conductors.

- Simplification passes over all BSDF models, reduced the number
  of unnecessary masked assignments.

- BSDFSample interface update: removed 'wi' field (already provided in
  SurfaceInteraction)

- BSDF interface change: BSDF context is now first argument of
  sample()/eval()/pdf() functions

- minor improvements in Frame::sin_phi, etc. helper functions

- disable Conda cache (AppVeyor)

- additional microfacet Python testcases

+++ src/librender/integrator.cpp
@@ -108,8 +108,8 @@ void SamplingIntegrator::render_block_scalar(const Scene *scene,
                                              ImageBlock *block) const {
     const Sensor *sensor = scene->sensor();
     block->clear();
-    uint32_t pixel_count = m_block_size * m_block_size,
-             sample_count = sampler->sample_count();
+    uint32_t pixel_count = (uint32_t) (m_block_size * m_block_size),
+             sample_count = (uint32_t) sampler->sample_count();
 
     RadianceSample3f rs(scene, sampler);
     bool needs_aperture_sample = sensor->needs_aperture_sample();
@@ -166,7 +166,7 @@ void SamplingIntegrator::render_block_vector(const Scene *scene,
                                              Point2fX &points) const {
     const Sensor *sensor = scene->sensor();
     block->clear();
-    uint32_t sample_count = sampler->sample_count();
+    uint32_t sample_count = (uint32_t) sampler->sample_count();
 
     if (block->size().x() != block->size().y()
         || (uint32_t) hprod(block->size()) * sample_count
